Fix cheesecake pancakes image directory
The image for the cheesecake pancakes recipe was not showing. This is
because initially the images were not located in their own
directory but directly under the recipes directory.

In this fix the src for the img tag in the cheesecake-pancakes.html
file was modified from "cheesecake-pancakes-complete.webp" to
"./images/cheesecake-pancakes-complete.webp".